Generate coverage data and prepare integration with Coveralls.io
I've updated the Travis build file and the Tox config to run nosetests
with the --with-coverage parameter. The resulting coverage information
can then be automatically submitted to Coveralls.io after each Travis
build (via the coveralls module).

I've also embedded a coverage badge to README.md so that we get a nice
coverage status once the repo is connected to Coveralls.